Add SchemaBuilder API
Wrap all methods of org.apache.avro.SchemaBuilder and its nested classes. Reimplements the full SchemaBuilder API, with the exceptions of the self methods and the shortcut methods of the nested classes. Side effects are now wrapped in cats.effect.Sync, and now typeclass composition replaces inheritance, but extension methods are used to remain faithful to the Java API.